Fixy wixy

+++ src/main/java/modtweaker/mods/bloodmagic/handlers/Alchemy.java
@@ -17,7 +17,7 @@ public class Alchemy {
     //Adding a Blood Magic Alchemical Chemistry Set recipe
     @ZenMethod
     public static void addRecipe(IItemStack output, IItemStack[] input, int tier, int lp) {
-        MineTweakerAPI.tweaker.apply(new Add(new AlchemyRecipe(toStack(output), lp / 100, toStacks(input), tier)));
+        MineTweakerAPI.tweaker.apply(new Add(new AlchemyRecipe(toStack(output), (int) (((double)lp) / 100), toStacks(input), tier)));
     }
 
     //Passes the list to the base list implementation, and adds the recipe
